threekingdomsmanualdownloadbug
anytime myself, a premium user attempts to manually download a mod of three kingdoms I get sent instead to a url that has tons of random symbols and no pop up to download the mod. tested on skyrim and fallout 4. does not do the same as three kingdoms. all was done on Microsoft edge